Add example for service-account,yaml
The service-account.yaml file is referred to but there is no mention of what it should actually contain. This is not very complex, but it very much breaks the flow of the install and causes you to have to break out and look up the correct syntax for a service account, whether it requires RBAC etc. 

Added a link to the RBAC notes also to help on that front.