docs(monitoring): replace monitoring-daemon with the observability-plugin
The monitoring-daemon was deprecated as of OSS 1.20 and was replaced with the OSS Observability Plugin,
this change updates the documentation to reflect this.

I personally spent a lot of time setting up the monitoring-daemon to only learn it was no longer supported so
this is an attempt to update the docs for everyone else :).

For the actual plugin configuration I'm linking to the official docs which I hope is okay.