fix(permissions): Pipeline permissions do not affect manual triggers (#444)
The documentation page for Pipeline Permissions explains that when roles
are added to a pipeline definition, only users with all of these roles
can then edit or execute the pipeline. This is true about edits, since
adding a privileged role there would let a user trigger an execution
with privileges they don't have, but this feature does not affect manual
executions. These still run with the roles and privileges of the user
who triggered the pipeline manually, regardless of Pipeline Permissions.